openclaw/requirements.txt
# Core AI
google-genai>=1.64.0
anthropic>=0.83.0
requests>=2.32.0
python-dotenv>=1.0.0

# Notifications
twilio>=9.0.0

# Testing
pytest>=8.0.0
pytest-cov>=5.0.0
